ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
tensorflow-cpu==2.13.1
numpy==1.24.4
python-dotenv==1.0.1
